Create @cached_property
The @cached_property decorator is similar to the @property decorator,
except that the function is only called to compute the value if it is
not already cached.  This cache can be controlled by using the value
of other attributes or properties, either of the object itself, or of
an attribute of that object.